refactor(tests): extract shared read/frontmatter helper
The three migrated test files each inlined identical read() and
frontmatter() helpers. Pull them into tests/helpers/text.ts and import
from all three, applying the Rule of Three. No assertion logic changes.
