+++ tests/protocol.test.ts
@@ -1,29 +1,10 @@
 import { describe, expect, test } from "bun:test";
-import { existsSync, readFileSync } from "node:fs";
+import { existsSync } from "node:fs";
 import { join } from "node:path";
 
+import { frontmatter, read } from "./helpers/text";
+
 const REPO_ROOT = process.cwd();
-
-function read(path: string): string {
-  return readFileSync(path, "utf8");
-}
-
-// Frontmatter slice: the lines strictly between the first and second `---`
-// markers. If fewer than two markers exist, the slice is empty ("") and
-// dependent assertions must fail, not skip.
-function frontmatter(text: string): string {
-  const lines = text.split("\n");
-  let count = 0;
-  const out: string[] = [];
-  for (const line of lines) {
-    if (/^---$/.test(line)) {
-      count++;
-      continue;
-    }
-    if (count === 1) out.push(line);
-  }
-  return out.join("\n");
-}
 
 // Flatten newlines so multi-line prose can be matched in one regex.
 function flat(text: string): string {
